# Calculate virtual_path once for all instances