--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,5 +1,9 @@
 # master
 
+* Calculate virtual_path once for all instances of a component class to improve performance.
+
+    *Brad Parker*
+
 # 2.17.1
 
 * Fix bug where rendering Slot with empty block resulted in error.