I don't think this will vary per instance but I'm really not familiar
with Rails' view rendering.

My, possibly flawed, micro-benchmarking suggests this'll get
ViewComponent a 5 time speed up. Those same benchmarks suggest this
takes ViewComponent from being a bit faster than partials to being much
faster :D. **N.B** This'll only have an impact when a view is rendering
many instances of the same component. Though now I say that ... it
might have an impact over time when class caching is turned on
in production mode, I haven't tested that.

Setup
=====

This is a sketch of what we've been using to bench:

    require "benchmark/ips"

    class ViewComponentPerformanceRenderer
      def initialize
        @controller = ViewComponent::Base.test_controller.constantize.new
        @controller.request = ActionDispatch::TestRequest.create
        @controller.extend(Rails.application.routes.url_helpers)
      end

      def render_komponent(name, **args, &block)
        @controller.helpers.component(name, args, &block)
      end

      def render_inline(component, **args, &block)
        @controller.view_context.render(component, args, &block)
      end
    end

    renderer = ViewComponentPerformanceRenderer.new

    Benchmark.ips do |benchmark|
      benchmark.report("Partial") do
        renderer.render_inline("shared/dummy") do
          "Hey, there"
        end
      end

      benchmark.report("ViewComponent") do
        renderer.render_inline(DummyComponent.new) do
          "Hey, there"
        end
      end

      benchmark.report("Komponent") do
        renderer.render_komponent("dummy") do
          "Hey, there"
        end
      end
    end

I've tried to make the components almost as simple as possible (they all
accept content so I could test the effect of nesting).

**N.B** I also enabled template caching but made sure that each example
still had an ERB template (E.G. no #call components).

Partial
=======

views/shared/_dummy.html.erb

    <div>
      <%= yield %>
    </div>

Komponent
=========

frontend/components/dummy/dummy_component.rb

    module DummyComponent
      extend ComponentHelper
    end

frontend/components/dummy/_dummy.html.erb

    <div>
      <%= yield %>
    </div>

ViewComponent
=============

app/components/dummy_component.rb

    class DummyComponent < ViewComponent::Base
    end

app/components/dummy_component.html.erb

    <div>
      <%= content %>
    </div>

Results
=======

Without this change:

    Warming up --------------------------------------
                 Partial   950.000  i/100ms
           ViewComponent     1.344k i/100ms
               Komponent   440.000  i/100ms
    Calculating -------------------------------------
                 Partial     11.522k (± 9.0%) i/s -     57.950k in   5.071796s
           ViewComponent     13.717k (± 6.7%) i/s -     68.544k in   5.021419s
               Komponent      4.442k (± 7.0%) i/s -     22.440k in   5.075137s

And with:

    Warming up --------------------------------------
                 Partial   946.000  i/100ms
           ViewComponent     5.731k i/100ms
               Komponent   431.000  i/100ms
    Calculating -------------------------------------
                 Partial     11.339k (± 9.2%) i/s -     56.760k in   5.048910s
           ViewComponent     56.099k (± 5.4%) i/s -    280.819k in   5.019196s
               Komponent      4.468k (± 7.3%) i/s -     22.412k in   5.041259s

---

I suspect that this is good for our ([Five Good Friends](https://github.com/fivegoodfriends/)) use-case, we
want to use lots of little ViewComponents as a way to build out a bit
of a component library. We'd like to have a few tiny components that
handle things like error states, laying things out in nicely spaced
rows, consistent border styles, spacing etc. This means that a few
basic components are used to build out the more complicated ones, and
those basic components get rendered a lot per view.