@bradparker 2020-07-22
Done and done :), I'm keen to add the bench-marking stuff too and will aim to do that soon in a follow up PR.